Refactoring: Package by feature walking skeleton  (#44)
