--- bot/commands.md
@@ -1,24 +1,0 @@
-# Commands
-
-## Key 
-| Symbol      | Meaning                        |
-|-------------|--------------------------------|
-| [Argument]  | Argument is not required.      |
-
-## Demo
-| Commands | Arguments     | Description               |
-|----------|---------------|---------------------------|
-| Add      | First, Second | Add two numbers together. |
-| run      |               | 開始遊戲                      |
-
-## Utility
-| Commands | Arguments | Description          |
-|----------|-----------|----------------------|
-| Help     | [Command] | Display a help menu. |
-| info     |           | Bot info for null    |
-
-## gamification
-| Commands     | Arguments | Description               |
-|--------------|-----------|---------------------------|
-| gamification |           | A 'gamification' command. |
-
