+++ commands.md
@@ -5,27 +5,14 @@
 |-------------|--------------------------------|
 | [Argument]  | Argument is not required.      |
 
-## Demo
-| Commands | Arguments     | Description               |
-|----------|---------------|---------------------------|
-| Add      | First, Second | Add two numbers together. |
-| run      |               | 開始遊戲                      |
-
 ## Utility
 | Commands | Arguments | Description          |
 |----------|-----------|----------------------|
 | Help     | [Command] | Display a help menu. |
 | info     |           | Bot info for null    |
 
-## gamification
-| Commands     | Arguments | Description               |
-|--------------|-----------|---------------------------|
-| gamification |           | A 'gamification' command. |
-
 ## two-sum
-| Commands    | Arguments     | Description                |
-|-------------|---------------|----------------------------|
-| mongo-query | id            | query something from mongo |
-| mongo-save  | note          | save something into mongo  |
-| two-sum     | First, Second | Add two numbers together.  |
+| Commands | Arguments     | Description               |
+|----------|---------------|---------------------------|
+| two-sum  | First, Second | Add two numbers together. |
