+++ docs/discord-bot-commands.md
@@ -1,0 +1,31 @@
+# Commands
+
+## Key 
+| Symbol      | Meaning                        |
+|-------------|--------------------------------|
+| [Argument]  | Argument is not required.      |
+
+## Demo
+| Commands | Arguments     | Description               |
+|----------|---------------|---------------------------|
+| Add      | First, Second | Add two numbers together. |
+| run      |               | 開始遊戲                      |
+
+## Utility
+| Commands | Arguments | Description          |
+|----------|-----------|----------------------|
+| Help     | [Command] | Display a help menu. |
+| info     |           | Bot info for null    |
+
+## gamification
+| Commands     | Arguments | Description               |
+|--------------|-----------|---------------------------|
+| gamification |           | A 'gamification' command. |
+
+## two-sum
+| Commands    | Arguments     | Description                |
+|-------------|---------------|----------------------------|
+| mongo-query | id            | query something from mongo |
+| mongo-save  | note          | save something into mongo  |
+| two-sum     | First, Second | Add two numbers together.  |
+
