--- domain/src/main/kotlin/tw/waterballsa/alpha/wsabot/domain/TwoSum.kt
@@ -1,7 +1,0 @@
-package tw.waterballsa.alpha.wsabot.domain
-
-class TwoSum(var a: Int, var b: Int) {
-    fun sum(username: String): String {
-        return "$username \n$a + $b = ${a + b}"
-    }
-}
