+++ main/src/main/resources/wsa.beta.properties
@@ -1,0 +1,5 @@
+guild-id=1038654765896310804
+unlock-entry-message-id=1038667013259792455
+self-intro-channel-id=1039196068455399474
+wsa-guest-role-id=1040635429541646407
+wsa-citizen-role-id=1038661590985228298
