+++ main/src/main/resources/wsa.prod.properties
@@ -1,0 +1,5 @@
+guild.id=937992003415838761
+unlock-entry-message-id=1042776102734663740
+self-intro-channel-id=937992281837961257
+wsa-guest-role-id=1042770137742319636
+wsa-citizen-role-id=954669665588756480
