+++ two-sum/src/main/kotlin/tw/waterballsa/utopia/twosum/domain/TwoSum.kt
@@ -1,0 +1,7 @@
+package tw.waterballsa.utopia.twosum.domain
+
+class TwoSum(var a: Int, var b: Int) {
+    fun sum(username: String): String {
+        return "$username \n$a + $b = ${a + b}"
+    }
+}
